+++ test/language.js
@@ -283,6 +283,12 @@ describe('negotiator.languages()', function () {
     })
   })
 
+  whenAcceptLanguage('en-US;foo=bar;q=1, en-GB;q=1', function () {
+    it('should return en-US, en-GB', function () {
+      assert.deepEqual(this.negotiator.languages(), ['en-US', 'en-GB'])
+    })
+  })
+
   whenAcceptLanguage('nl;q=0.5, fr, de, en, it, es, pt, no, se, fi, ro', function () {
     it('should use prefer fr over nl', function () {
       assert.deepEqual(this.negotiator.languages(), ['fr', 'de', 'en', 'it', 'es', 'pt', 'no', 'se', 'fi', 'ro', 'nl'])